mxshop-manage/mxshop_srvs/pyproject.toml
[project]
name = "mxshop-srvs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "alibabacloud-credentials==1.0.2",
    "alibabacloud-credentials-api==1.0.0",
    "alibabacloud-darabonba-array==0.1.0",
    "alibabacloud-darabonba-encode-util==0.0.2",
    "alibabacloud-darabonba-map==0.0.1",
    "alibabacloud-darabonba-signature-util==0.0.4",
    "alibabacloud-darabonba-string==0.0.4",
    "alibabacloud-endpoint-util==0.0.4",
    "alibabacloud-gateway-pop==0.0.9",
    "alibabacloud-gateway-spi==0.0.3",
    "alibabacloud-kms20160120==2.3.1",
    "alibabacloud-openapi-util==0.2.2",
    "alibabacloud-tea==0.4.3",
    "alibabacloud-tea-openapi==0.4.0",
    "alibabacloud-tea-util==0.3.13",
    "annotated-types==0.7.0",
    "apscheduler==3.11.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "cryptography==44.0.3",
    "darabonba-core==1.0.2",
    "faker==37.5.3",
    "frozenlist==1.7.0",
    "grpcio==1.74.0",
    "grpcio-tools==1.74.0",
    "idna==3.10",
    "jaeger-client==4.8.0",
    "loguru==0.7.3",
    "multidict==6.6.3",
    "mypy-protobuf>=3.6.0",
    "nacos-sdk-python==2.0.8",
    "opentracing==2.4.0",
    "passlib==1.7.4",
    "peewee==3.18.2",
    "pool>=0.1.2.dev0",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pycparser==2.22",
    "pycryptodome==3.23.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pymysql==1.1.1",
    "python-consul2==0.1.5",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.4",
    "setuptools==80.9.0",
    "shortcuts>=0.11.0",
    "six==1.17.0",
    "threadloop==1.0.2",
    "thrift==0.22.0",
    "tornado==6.5.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "yarl==1.20.1",
]
